+++ cryptofeed/rest/coinbase.py
@@ -145,7 +145,7 @@ class Coinbase(API):
     def _trade_normalize(self, symbol: str, data: dict) -> dict:
         return {
             'timestamp': pd.Timestamp(data['time']).timestamp(),
-            'pair': symbol,
+            'symbol': symbol,
             'id': data['trade_id'],
             'feed': self.ID,
             'side': SELL if data['side'] == 'buy' else BUY,
@@ -196,7 +196,7 @@ class Coinbase(API):
         data = self._request('GET', f'/products/{symbol}/ticker', retry=retry, retry_wait=retry_wait)
         self._handle_error(data, LOG)
         data = data.json()
-        return {'pair': symbol,
+        return {'symbol': symbol,
                 'feed': self.ID,
                 'bid': Decimal(data['bid']),
                 'ask': Decimal(data['ask'])