+++ cryptofeed/rest/rest.py
@@ -17,7 +17,7 @@ from cryptofeed.rest.ftx import FTX
 from cryptofeed.rest.gemini import Gemini
 from cryptofeed.rest.kraken import Kraken
 from cryptofeed.rest.poloniex import Poloniex
-from cryptofeed.standards import load_exchange_pair_mapping
+from cryptofeed.standards import load_exchange_symbol_mapping
 
 
 LOG = logging.getLogger('rest')
@@ -36,7 +36,7 @@ class Rest:
     """
 
     def __init__(self, config=None, sandbox=False):
-        self.config = Config(file_name=config)
+        self.config = Config(config=config)
 
         lfile = 'rest.log' if not self.config or not self.config.restlog.filename else self.config.restlog.filename
         level = logging.WARNING if not self.config or not self.config.restlog.level else self.config.restlog.level
@@ -59,9 +59,9 @@ class Rest:
         exch = self.lookup[key.lower()]
         if not exch.mapped:
             try:
-                load_exchange_pair_mapping(exch.ID + 'REST')
+                load_exchange_symbol_mapping(exch.ID + 'REST')
             except KeyError:
-                load_exchange_pair_mapping(exch.ID)
+                load_exchange_symbol_mapping(exch.ID)
             exch.mapped = True
         return exch
 
@@ -69,8 +69,8 @@ class Rest:
         exch = self.lookup[attr.lower()]
         if not exch.mapped:
             try:
-                load_exchange_pair_mapping(exch.ID + 'REST')
+                load_exchange_symbol_mapping(exch.ID + 'REST')
             except KeyError:
-                load_exchange_pair_mapping(exch.ID)
+                load_exchange_symbol_mapping(exch.ID)
             exch.mapped = True
         return exch