+++ CHANGES.md
@@ -12,6 +12,10 @@
   * Bugfix: wrong datatype in BackendFuturesIndexCallback
   * Bugfix: fix bad postgres callback for open_interest and futures_index
   * Feature: Signal handler installation now optional, can be done separately. This will allow the feedhandler to be run from child threads/loops
+  * Bugfix: fix binance delivery book ticker (message format change)
+  * Breaking change: Feed object `config` renamed `subscription`
+  * Feature: Configuration passed from feedhandler to exchanges
+  * Breaking change: most use of `pair` and `pairs` changed to `symbol` and `symbols` to be more consistent with actual usage. pairs.py renamed to symbols.py
 
 ### 1.6.2 (2020-12-25)
   * Feature: Support for Coingecko aggregated data per coin, to be used with a new data channel 'profile'